Allow playing output to default output device when during uplink playback
Change the current behavior of AudioPolicyManager during incall music playback to allow apps that want to play audio with AUDIO_STREAM_MUSIC to default device to route the audio to the default device chosen by the engine (and not the one forced by the ongoing incall music routing). The current behavior still playes the audio to the default output device, which will be set to the TELEPHONY_TX device whenever there is ongoing uplink playback. This change will only affect a case in which uplink playback is in progress and another app tries to play audio using the music stream at that time.

Test: Tested manually that the behavior described in the bug is fixed and that both apps can play audio to two different output devices.
Bug: 111467967.
Change-Id: I25bccb635ea436bb6827483732e53c49c4b27825