MediaCas: return audio samples in scrambled form
We can no longer assume audio stream can always be
descrambled into clear buffers. If audio is scrambled,
handle it the same way and extract samples in scrambled
form. Set a dummy format on the track to provide info
about stream type only, and leave it to the decoder
to handle the format change.

bug: 111271614
Test: CTS MediaDrmClearkeyTest#testClearKeyPlaybackMpeg2ts
MediaDrmClearKeyTest#testPlaybackMpeg2ts
Change-Id: I2a803a92f395f06af54f3ad65a5c436edd7ce251