Update patch set 3
Patch Set 3:

> How do you guarantee that an update of the third party cam app isnt
 > removing their shutter sound and due to tis change you end up with
 > no sound? Better make the third party apps behave properly than
 > hacking this

Typically, this is the stock camera. The functional does not change.

I do not understand why this can not be added, if at the moment the sound of the camera does not work in the "do not disturb" mode?
Turn on the "do not disturb" mode and any camera does work without sound.

Patch-set: 3
Reviewer: Shev T <9760@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>