NBLog: Restore log dump and add new logging types
Merging is temporarily disabled for now because the readers managed by
the merger are the same as the ones used for the dumpsys log dump, and
reading from the log effectively consumes the buffer. Eventually, the
readers for the two functionalities will need to be separated to avoid
conflict. The dump of the merged buffer is also disabled, which removes
the dumping of histograms.

The new types added are monotonic thread cycle time, CPU thread cycle
time, CPU frequency, and latency (which will need to be specified in
further detail later). Logging support is added only for monotonic
thread cycle time at the moment.

Test: build, log monotonic thread cycle time in FastThread, check output
in dumpsys media.log -r.

Change-Id: I1b781d6db102fb917fd0bac964eeebd0309234c0
Bug: 68148948