Update patch set 7
Patch Set 7:

> Yup, I use it.
 > 
 > https://github.com/LineageOS/android_device_samsung_msm8930-common/commit/fecb8efa87a2a3c8300dcb35fd2ee754b61537f2

Interesting. This is not the original change. I had to adapt it to make it compatible with the new go modules, but I was not sure if it was correct because it did not work during initial tests. I need to check again. In the meanwhile, I was going to rework the change in order to go with: https://review.lineageos.org/#/c/LineageOS/android_vendor_lineage/+/226156/

Patch-set: 7
Reviewer: Bruno Martins <13648@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>