Audioflinger: Dump "master mute" state in threads
"Master mute" which can be toggled on via "ro.audio.silent" property
has user-hearable effect, but is not reflected in the dumpsys.
This can cause confusion (see the bug referenced).

Note that we need to dump PlaybackThread::mMasterMute, not
AudioFlinger::mMasterMute, as the latter does not get updated
by threads in order to avoid taking a lock.

Bug: 113348789
Test: toggle property, check media.audio_flinger dump
Change-Id: I7f276b233df2f5eca585ffb30031835c030858e1