audio: fix AudioTrack and AudioRecord restore
Allow retry on failure to reconnected to audioserver when restoring an
AudioTrack or AudioRecord and do not change state to inactive when
reconnection fails. In case of failure due to temporary race conditions this will
increase the chances of success and avoid ending up in an inconsistent
active state if a later reconnection succeeds.

Bug: 113109529
Test: AudioTrack CTS tests and manual playback smoke tests.
Test: AudioRecord CTS tests and manual capture smoke tests.
Change-Id: I570724eb46f3a0f830267396aba06297bd37df7a