Add media.codec.update service
This is done similarly to how we update the extractors.
media.swcodec will start a service "media.codec.update"
to get notified of the update apk location, then updated
version of the component store will be loaded.

Since the update interface is exactly the same, we reuse
IMediaExtractorUpdateService to avoid adding new an aidl
interface. Besides this service will likely be removed
after we switch to APEX.)

Bug: 111407413
Bug: 117290290
Change-Id: I739ba0071806beed99302ccd3da7a6f73a437f06