audiopolicy: apm: updateCallRouting shall use SW Bridging if no external route support
While using Routing3.0 API, and XML configuration file, the route section
shall declare the topology supported by a given platform.
A Device sink to device source route declaration implies that an external
route is supported and AudioHAL can configure it.

In case of call, even if the Rx and Tx device belongs to respectively
primary output and input of the primary module, it does not mean AudioHAL
can connect the telephony Rx and Tx by itself unless a Route it declared.

This is the purpose of this patch to take benefit of XML file.

Test: make

Change-Id: Id62ab3889f20b4daacf7adcce4d7688caf509443
Signed-off-by: François Gaffie <francois.gaffie@renault.com>