audiopolicy: use an exact match for MMAP input
AudioFlinger was returning a stereo input if mono was requested
but not supported.
But AAudio did not know it was stereo so the data was corrupted.
So if MMAP input is requested, do an exact match in
isCompatibleProfile().

Bug: 119518992
Test: see bug for repo steps
Test: adb shell aaudio_loopback -tm -pl -Pl -x -X -C1
Change-Id: I925efecca9c99b8a2b7bde3826da4d9e9e663ee1
Merged-In: I925efecca9c99b8a2b7bde3826da4d9e9e663ee1