Fix NE in pause process when encoder has been released
When the encoder generates an error or the recording file reaches its maximum
size, the MediaRecorder would enter its stop process automatically and the encoder
would be released in this stop process. If user tries to pause the recorder when
the encoder has been released, the code would run into a wrong branch to
CHECK(mPuller!=NULL), and this would cause NE for there is no puller in SURFACE
mode. For we don't need to do anything in stop process within SURFACE mode when
the encoder has been released, this patch just adds an empty branch.

Bug: 127673985
Test: CTS - MediaRecorderTest
 ANDROID_BUILD_droid-cts/tools/cts-tradefed run cts -m
 CtsMediaTestCases --test android.media.cts.MediaRecorderTest

Change-Id: I6403855f1abb86a6b125fc5ba2c5e5c31a5241db