C2SoftVorbisDec: Ignore error from vorbis_dsp_synthesis
Error from vorbis_dsp_synthesis() is to be ignored.
After seek in some cases, vorbis_dsp_synthesis() returns
OV_ENOTAUDIO when input is not an audio packet.
This results in seek failing for vorbis files.

OMX plugin for Vorbis decoder ignores the errors returned from this
function. C2 plugin is updated to do the same.

Bug: 110759274
Test: cts-tradefed run commandAndExit cts-dev -m CtsMediaTestCases \
-t android.media.cts.DecoderTest

Change-Id: Ic98d0555121c7f741c468c69f1acdb75036d8c3d