Suppress a static-analyzer warning
Our analyzer was flagging this as a memory leak, since it saw that we
were `strdup`ing into the mData member of this AString temporary:

frameworks/av/media/libstagefright/foundation/AString.cpp:368:5:
warning: Potential leak of memory pointed to by field 'mData'
[clang-analyzer-unix.Malloc]

This is incorrect, and appears to be related to LLVM PR38176.

Add a NOLINT until it's fixed upstream.

Bug: 120788908
Test: Ran the analyzer
Change-Id: Iaee20a740d721010862d2cce51383bdae311bd0a