Enable Scudo for mediaswcodec
Scudo is a hardened usermode allocator that aims at mitigating
heap-based vulnerabilities (heap overflow, double free, etc).

See go/scudo-allocator for more information.

This change enables the allocator for the mediaswcodec service,
effectively linking in the Scudo dynamic library to the binary.
Allocation functions will be serviced by the allocator.

The observed impact on the memory footprint of the service is about a
500k increase to the RSS (due to the Quarantine and chunk headers
overhead), and the MediaCts tests pass without any issue I could see.

In the event of a heap bug detected, the allocator will abort the
process with a 'Scudo ERROR' message in the log followed by the cause
(corrupted header, invalid state & so on) and the stack trace.

Test: CtsMediaTestCases on a marlin
Bug: 63907455
Change-Id: I751af455c002a59ccc4d89cd1e29743036f27708