Audio HAL: Simplify types updates for 'common' and 'effect' packages
Instead of importing every single type individually ("using <type>"),
import entire namespaces of
::android::hardware::audio::common::CPP_VERSION and
::android::hardware::audio::effect::CPP_VERSION
inside the "implementation" namespace in the default implementation.

This simplifies modifying types, as it is not needed anymore
to surround "using <type>" directives with #if MAJOR_VERSION ...
Note that the contents of the namespaces are imported inside
the "implementation" namespace, which reduces risks of name collision

Enclosed EffectHalHidl and supporting classes into 'effect' namespace
because importing namespaces of ::a::h::audio::effect::CPP_VERSION and
::a::h::audio::CPP_VERSION into the same namespace, as it happens
for StreamHalHidl.cpp module, causes name conflicts.

Test: make
Change-Id: Ia37d672605c7541b19ee40fc1b58eabaa9ed8e17