Enable Scudo for mediaextractor
Scudo is a hardened usermode allocator that aims at mitigating
heap-based vulnerabilities (heap overflow, double free, etc).

See go/scudo-allocator for more information.

This change enables the allocator for the mediaextractor service,
effectively linking in the Scudo dynamic library to the binary.
Allocation functions will be serviced by the allocator.

The 'deallocation-type-mismatch' check had to be disabled as some
third party libraries were triggering it.

In the event of a heap bug detected, the allocator will abort the
process with a 'Scudo ERROR' message in the log followed by the cause
(corrupted header, invalid state & so on) and the stack trace.

Test: CtsMediaTestCases on a marlin
Bug: 63907455
Change-Id: I9e0ea03c0d39e723a880e7fcd761fe0718bbbacd