AudioPolicy: enable apps to mute BLUETOOTH_SCO
This change enables an app that has the MODIFY_PHONE_STATE
permission to mute the BULETOOTH_SCO stream.
The change adds specific handling in AudioPolicyManager
for the case in which volume index is 0 and stream type is
BLUETOOTH_SCO.
The permission check is done in the AudioService Java code for the
calling process.

Bug: 111195322
Test: Tested manually
Change-Id: I5085826a7287dfbc2cfb17a9bd2d3ca1af66ea4a