stagefright: Updates to Opus unified CSD syntax
Added a marker and length field before OpusHead data as well.
This will prevent any potential truncation of OpusHead, when
data inside OpusHead matches one of the markers defined

Added checks to validate sizes parsed.
If the sizes are not as expected, then treat that as an error

OggWriter updated to not write codec config data as frame data

Test: With a local AMediaCodec api based application to decode
 opus encoder's output
Test: Test with mediamuxer api based application to mux encoders output
 to ogg file
Test: cts-tradefed run cts -m CtsMediaTestCases \
 -t android.media.cts.EncoderTest#testOpusEncoders

Bug: 123581317
Bug: 124053011
Change-Id: Ic3c7613ff47855e16be39dc60939e1e715522bc6