MKV supports MP3 audio whose CodecID is A_MS/ACM
CodecID for MKV MP3 audio is not only A_MPEG/L3, but also
AMS_ACM. When CodecID is AMS_ACM, we should check format ID and
the file is MP3 if format ID is 0x0055. Need to support this
MP3 type.

Bug: 123594233
Test: Test MKV file whose audio CodecID is AMS_ACM and format
ID is 0x0055, and check if the MP3 audio can be played.

Change-Id: I15129e276b19338dc4f3f09eee6020aca5a5fa8e