Mark functions in IMtpDatabase as pure virtual
Bug: http://b/116873221

Mark beginCopyObject() and endCopyObject() as pure virtual.  If not,
Clang assumes that the TU which defines this function will contain the
vtable for the class as well.  Since there's no out-of-line definition
of this funciton, no TU ends up with the vtable.

This causes a problem with coverage builds, which are built with -O0,
where calls don't get inlined, thereby requiring a definition of the
vtable.  For non -O0 builds, the vtable is not required since the
virtual calls get inlined/optimized-out.

Test: Build with and without -O0.
Change-Id: I5e8dcf583a75af51af1e9dd31a6635aa2f9d4413