Mark functions in MediaPlayerBase::AudioSink as pure virtual
Bug: http://b/116873221

Mark the functions in this class that aren't defined as pure virtual.
If not, Clang assumes that the TU which defines the first undefined
function will contain the vtable for the class as well.  Since there's
no out-of-line definition of this funciton, no TU ends up with the
vtable.

This causes a problem with coverage builds, which are built with -O0,
where calls don't get inlined, thereby requiring a definition of the
vtable.  For non -O0 builds, the vtable is not required since the
virtual calls get inlined/optimized-out.

Test: Build with and without coverage
Change-Id: I4644dcd892c5f47031f1603b0f809c1da95314db