Cut dependency to libandroid.so
This is a part of job to cut dependency from libmediandk
to libandroid_runtime. As a LL-NDK, libmediandk desn't have
to depend on libandroid_runtime.

depependcy:
* libmediandk -> libandroid -> libandroid_runtime.

Libmediandk uses only one symbol(ANativeWindow_aquire) from
libandroid and it is re-exported from libnativewindow.

So by switching from libandroid to libnativewindow, the dependency
from libmediandk to libandroid is cut easily.

Bug: 124268753
Test: m -j
Change-Id: I5da320780dba7c63ce39679dbf5d40fe9202eb00