AMRWB_E_SAMPLE: Fix resource leaks in error cases
We also avoid a pontential call to dlclose(NULL) which was
possible in the previous code.

Note the leaks aren't really consequential since this is a
standalone sample.  But by putting in simple fixes for these,
we reduce the noise when using static analysis to find leaks
in Android code.

Test: TreeHugger

Change-Id: I3f69e410ecf611dcb60e4f58aae27801b2141e4d