StagefrightRecorder:close next filedescriptor
Bug: 127693962

StagefrightRecorder receives next file descriptor in it's function setNextOutputFile() from
MediaRecorder.setNextOutputFile(). It duplicates the received file descriptor, but doesn't close
the duplicated one when done.

This CL fixes that issue by not duplicating it in StagefrighRecorder as it has to just pass it on
to Mpeg4Writer.  Mpeg4Writer duplicates received file descriptor.  It frees them during release.

Test: 	cts-tradefed run cts-dev -m CtsMediaTestCases --compatibility:module-arg
	CtsMediaTestCases:include-annotation:android.platform.test.annotations.RequiresDevice

Change-Id: I24fc12bd957a2abdb8785625261c95da62952736