Make NuMediaExtractor behave more like mediaextractor
NuMediaExtractor does limited read-ahead, then fails the current
read immediately if an error was encountered during the read-ahead.
This prevents reading the first buffer when there's an error reading
the second buffer, making it impossible to reproduce some behaviors
that *can* be reproduced with stagefright or by talking to the
service directly.
This change make it behave more like the underlying mediaextractor.

In addition, this fixes the behavior of advance(), which is supposed
to return false if no more samples are left. Prior to this change,
advance would return true, but then the next read would fail.

Bug: 121204004
Test: manual, CTS

Change-Id: I8f906945713717ee6be9dff30165635e03389153