Add /system/lib to the search paths for asan
When running with asan, the platform namespace didn't have /system/lib,
but /data/asan/system/lib in its search paths. As a result, clang
runtime libraries like libclang-rt.asan-i686-android.so which are still
in /system/lib cound not be found.

Fixing the problem by adding /system/lib to the search paths after the
asan path.

Bug: 131170196
Test: check if mediaswcodec process is running in aosp_cf_x86_pasan

Change-Id: Iffe5d25613999a00ead5eb03c0efca1d6c9ce012