Install extractors for 1st and 2nd ABIs to the test APEX
The test APEX is built with aosp_x86_64 where first ABI is 64-bit.
But the test APEX can be installed to 32-bit target like cf_x86_phone.
The lack of 32-bit extractor libraries in the 32-bit target can cause
problem. So, install both ABIs of the extractor libraries to the test
APEX.

Bug: 125914810
Test: choosecombo to aosp_x86_64 and build test_com.android.media
Check that the test APEX has both lib and lib64 directories.

Change-Id: Ib073f61838fea0b823c870c8c71b49a7ecc9346e