Do Not Pass Blank App Package Name to MediaDrm Plugins
Previously, the NDK has always passed a blank app package name to the
MediaDrm plugins. However, some DRM systems need this value to be filled
in to operate correctly.

Unfortunately, since supplying or not supplying this value can
invalidate existing DRM licenses, we cannot start passing it in for
devices that might have previously run without it. As such, passing it
is only supported on devices for which Q is their first Android release.

Bug: 65680731
Test: CTS NativeMediaDrmClearkeyTest
Change-Id: Icccde9708766a86ccd352c82be25285fc9f8ebca